Refactor events into event handler
- Add event loader to index.js
- Move event files into ./events
- Add informational events into index.js
- Add logging for event and command loader
- Add eslint warning for missing semicolons to increase consistency
- Lint all files and apply linting fixes
